ANE spike findings (docs/coreml-tradeoffs.md):
- CpuAndGpu is the correct Core ML compute unit for all current models
  (all VLMs and Qwen3-ASR exceed the ANE-sweet-spot threshold)
- Async-overlap vision carve-out closed for Gemma4 E4B; infrastructure retained
- Qwen3-ASR streaming overlap unconditionally free at any real speech density
- gemma4-mlx: fix quantized VL weight loading for E4B/12B (key normalization
  + MaybeQuantized<Linear> for embed_vision projection)

Also: systematic workspace hardening review across all crates.

# Conflicts:
#	gpt-sovits-mlx/src/models/t2s.rs
#	mlx-rs-core/src/lib.rs
#	mlx-rs-core/src/memory.rs
#	mlx-rs-core/src/metal_kernels.rs
#	qwen-image-mlx/src/pipeline.rs
#	qwen-image-mlx/src/qwen_quantized.rs
#	qwen3-mlx/src/model.rs

+++ gpt-sovits-mlx/src/nn/weight_norm.rs
@@ -145,9 +145,9 @@ impl WeightNormConv1d {
         dilation: i32,
     ) -> Result<Self, Exception> {
         let shape = weight_v.shape();
-        let out_channels = shape[0] as i32;
-        let kernel_size = shape[1] as i32;
-        let in_channels = shape[2] as i32;
+        let out_channels = shape[0];
+        let kernel_size = shape[1];
+        let in_channels = shape[2];
 
         Ok(Self {
             weight_g: Param::new(weight_g),
@@ -320,9 +320,9 @@ impl WeightNormConvTranspose1d {
         padding: i32,
     ) -> Result<Self, Exception> {
         let shape = weight_v.shape();
-        let out_channels = shape[0] as i32;
-        let kernel_size = shape[1] as i32;
-        let in_channels = shape[2] as i32;
+        let out_channels = shape[0];
+        let kernel_size = shape[1];
+        let in_channels = shape[2];
 
         Ok(Self {
             weight_g: Param::new(weight_g),
@@ -394,6 +394,7 @@ impl WeightNormConvTranspose1d {
     /// **Important**: We don't scale g because:
     /// - Before normalize: weight = g * v / ||v||
     /// - After normalize:  weight = g * (v/||v||) / 1 = g * v / ||v|| (same!)
+    ///
     /// So the computed weight is preserved when we don't scale g.
     pub fn normalize_v(&mut self) -> Result<(), Exception> {
         // Compute current norm over out_channels and kernel dims (axes 0, 1)
